chore(maintainability): Rename variables at to not hide class fields
